Remove AssignDefaults from TEdgeConfigDto
- We handle this within TEdgeConfig
