+++ tedge_config/src/providers/tedge_config_dto/mod.rs
@@ -1,6 +1,7 @@
 use crate::*;
 use serde::{Deserialize, Serialize};
-use std::path::PathBuf;
+
+const DEFAULT_ROOT_CERT_PATH: &str = "/etc/ssl/certs";
 
 /// Represents the complete configuration of a thin edge device.
 /// This configuration is a wrapper over the device specific configurations
@@ -80,70 +81,6 @@ pub(crate) struct AzureConfigDto {
     pub(crate) root_cert_path: Option<String>,
 }
 
-
-const DEVICE_KEY_FILE: &str = "tedge-private-key.pem";
-const DEVICE_CERT_FILE: &str = "tedge-certificate.pem";
-const TEDGE_HOME_DIR: &str = ".tedge";
-
-// XXX: Move to TEdgeConfig
-pub trait AssignDefaults: Sized {
-    fn assign_defaults(self) -> Result<Self, ConfigError>;
-}
-
-impl AssignDefaults for TEdgeConfigDto {
-    fn assign_defaults(self) -> Result<Self, ConfigError> {
-        let device_config = self.device.assign_defaults()?;
-
-        Ok(TEdgeConfigDto {
-            device: device_config,
-            ..self
-        })
-    }
-}
-
-impl AssignDefaults for DeviceConfigDto {
-    fn assign_defaults(self) -> Result<Self, ConfigError> {
-        let key_path = match self.key_path {
-            None => default_device_key_path()?,
-            Some(val) => val,
-        };
-
-        let cert_path = match self.cert_path {
-            None => default_device_cert_path()?,
-            Some(val) => val,
-        };
-
-        Ok(DeviceConfigDto {
-            key_path: Some(key_path),
-            cert_path: Some(cert_path),
-            ..self
-        })
-    }
-}
-
-pub(crate) fn default_device_key_path() -> Result<String, ConfigError> {
-    path_in_cert_directory(DEVICE_KEY_FILE)
-}
-
-pub(crate) fn default_device_cert_path() -> Result<String, ConfigError> {
-    path_in_cert_directory(DEVICE_CERT_FILE)
-}
-
-fn path_in_cert_directory(file_name: &str) -> Result<String, ConfigError> {
-    home_dir()?
-        .join(TEDGE_HOME_DIR)
-        .join(file_name)
-        .to_str()
-        .map(|s| s.into())
-        .ok_or(ConfigError::InvalidCharacterInHomeDirectoryPath)
-}
-
-fn home_dir() -> Result<PathBuf, ConfigError> {
-    // The usage of this deprecated method is temporary as this whole function will be replaced with the util function being added in CIT-137.
-    #![allow(deprecated)]
-    std::env::home_dir().ok_or(ConfigError::HomeDirectoryNotFound)
-}
-
 impl QuerySetting<AzureUrlSetting> for TEdgeConfigDto {
     fn query(&self, _setting: AzureUrlSetting) -> ConfigSettingResult<ConnectUrl> {
         self.azure
@@ -220,8 +157,6 @@ impl QuerySetting<C8yRootCertPathSetting> for TEdgeConfigDto {
             })
     }
 }
-
-const DEFAULT_ROOT_CERT_PATH: &str = "/etc/ssl/certs";
 
 impl QuerySettingWithDefault<AzureRootCertPathSetting> for TEdgeConfigDto {
     fn query_with_default(
